Change `ulem` to `lua-ul`; refs #277
This hack just changes in the DATA section:

ulem -> lua-ul
[normalem] ->
\sout -> \strikeThrough
\uwave -> \underLine

When this is used, output must be compiled with lualatex.

Some thoughts on implementing this in upstream latexdiff:

1. To make this an actual option, instead of hardcoding this change,
one would probably want to make general commands that call the right
ulem or lua-ul commnad, with separate preamble sections for either
ulem:

\RequirePackage[normalem]{ulem}
\newcommand[1]{\DIFstrikethrough}{\sout{#1}}
\newcommand[1]{\DIFunderline}{\uwave{#1}}

or lua-ul

\RequirePackage{lua-ul}
\newcommand[1]{\DIFstrikethrough}{\strikeThrough{#1}}
\newcommand[1]{\DIFunderline}{\underLine{#1}}

and then could make these changes instead

\RequirePackage[normalem]{ulem} ->
\sout -> \DIFstrikethrough
\uwave -> \underLine

This, however, would require adding in the logic to have either lua-ul or ulem into
the code.

I also hope the indirection like this won't break some LaTeX macro
somewhere but can't be sure of that.

Other options I considered would be:
- duplicating all these sections to have a ulem and a lua-ul
version (yuck)
- searching for these words and replacing them when reading in
preamble sections (D:)
- simply eliminating ulem support (working around it certainly seems to have caused a
lot of headaches for the creators of latexdiff)

2. lua-ul doesn't have the capacity out of the box to produce wavy
underlines like ulem, but is very configurable and it could be added.
Instructions are even in the lua-ul manual.